Assignment resubmission: Q6 fix
Corrects the segmentation fault by changing one line of HW3.cpp in
function histEq().  The change calls the push_back() method of the
std::vector class, instead of making assignment to a vector element
accessed via the [] operator.

This program also now reads the input image in full color to take
advantage of the color capability in the function histEq().  The
result of equalization is then converted back to greyscale.